+++ index.html
@@ -87,7 +87,7 @@
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="title">Заголовок объявления</label>
-          <input id="title" name="title" type="text" placeholder="Милая, уютная квартирка в центре Токио" minlength="30" maxlength="100" required
+          <input id="title" name="title" type="text" placeholder="Милая, уютная квартирка в центре Токио" minlength="30" data-pristine-maxlength="100" required
           data-pristine-required-message="Поле обязательно для заполнения"
           data-pristine-minlength-message="Минимальная длина 30 символов"
           data-pristine-maxlength-message="Максимальная длина 100 символов">
@@ -111,7 +111,7 @@
           <input id="price" name="price" type="number" placeholder="5000" min="0" max="100000" required
           data-pristine-required-message="Поле обязательно для заполнения"
           data-pristine-min-message="Минимальная цена за ночь 0 руб."
-          data-pristine-max-message="Максимальная цена за ночь 100.000 руб.">>
+          data-pristine-max-message="Максимальная цена за ночь 100.000 руб.">
           <div class="ad-form__slider"></div>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--time">
